+++ src/routes/envman/storage.ts
@@ -1,5 +1,6 @@
 import { Elysia } from 'elysia'
 import { storageCoreRouter } from './storage-core'
+import { storageFolderRouter } from './storage-folder'
 import { storageMoveRouter } from './storage-move'
 import { storageRenameRouter } from './storage-rename'
 import { storageUploadRouter } from './storage-upload'
@@ -9,3 +10,4 @@ export const storageRouter = new Elysia()
   .use(storageUploadRouter)
   .use(storageRenameRouter)
   .use(storageMoveRouter)
+  .use(storageFolderRouter)